mostly minor formatting fixes * removed unnecessary @EnableTransactionManagement from application * externalized feign config * moved Employee model into its own package * migrated properties to application.yml
